Fix compiler warning
src/winscard_msg.c:394:14: warning: implicit conversion loses integer precision: 'ssize_t' (aka 'long') to 'int' [-Wshorten-64-to-32]
                        written = send(filedes, buffer, remaining, MSG_NOSIGNAL);
                                ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/winscard_msg.c
@@ -378,7 +378,7 @@ INTERNAL LONG MessageSend(void *buffer_void, uint64_t buffer_size,
 		/* try to write only when the file descriptor is writable */
 		if (pollret > 0)
 		{
-			int written;
+			ssize_t written;
 
 			if (!(write_fd.revents & POLLOUT))
 			{
